chore: stub-ify the custom shader compiler example
