+++ Cargo.toml
@@ -740,8 +740,7 @@ flate2 = "1.0"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1.0.140"
 bytemuck = "1"
-# Needed for shader_material_hlsl example
-shaderc-dyn = "0.10.1"
+
 # The following explicit dependencies are needed for proc macros to work inside of examples as they are part of the bevy crate itself.
 bevy_animation = { path = "crates/bevy_animation", version = "0.19.0-dev", default-features = false }
 bevy_asset = { path = "crates/bevy_asset", version = "0.19.0-dev", default-features = false }
@@ -3082,14 +3081,14 @@ category = "Shaders"
 wasm = true
 
 [[example]]
-name = "shader_material_hlsl"
-path = "examples/shader/shader_material_hlsl.rs"
+name = "custom_shader_compiler"
+path = "examples/shader/custom_shader_compiler.rs"
 doc-scrape-examples = true
 required-features = ["shader_format_spirv"]
 
-[package.metadata.example.shader_material_hlsl]
-name = "Material - HLSL"
-description = "A material that uses an HLSL shader with a custom compiler"
+[package.metadata.example.custom_shader_compiler]
+name = "Custom shader compiler"
+description = "Shows how to setup a custom shader compiler"
 category = "Shaders"
 wasm = false
