+++ crates/bevy_render/src/render_resource/mod.rs
@@ -74,7 +74,6 @@ pub use naga::ShaderStage;
 
 pub use bevy_shader::{
     CompiledShader, NagaOilCompiler, ShaderCompileError, ShaderCompiler, ShaderLanguage,
-    SpirVPassthroughCompiler,
 };
 
 pub use bevy_material::{
